requirements.txt
colorama==0.4.6
pillow==12.0.0
tqdm==4.67.1
